client: refactor client to multiple files
