+++ renderapi/client/__init__.py
@@ -1,1 +1,4 @@
 from .client import *
+from .client_calls import *
+from .utils import *
+from .params import *
